Improvements to schema exports
Added a SchemaContainer that can hold multiple Schemas (sets of Schema
entitites).  This provides the means to store metadata for the
collection of Schema objects.

Added a SpatialReference attribute to the specification for Schemas and
Schema.

Commented out execution of the comparison tests and deleted the test
results.

Regenerated the Schemas.csv file.