Removed the status field from Cyber
Removed the status field from Cyberspace since it has the
operationalcondition field already.

Also added support for a DomainName attribute on Schema objects, so the
code that exports a Schema's SymbolSet Domain file can use a different
file name than the Fields schema file for the same Schema object.

Coded_Domain_Land.csv and Fields_LandEquipment.csv is the one example of
this difference in file name at the moment.